chore(logger): reorganise exports to remove circular dependency
