+++ src/logger/logLevel.ts
@@ -1,0 +1,7 @@
+export enum LogLevel {
+  Debug = 'Debug',
+  Info = 'Info',
+  Warn = 'Warn',
+  Error = 'Error',
+  Off = 'Off'
+}
